Rename rovr-generator to rovr-renderer
It makes a little more sense since this plugin only renders React
components and markdown code.
