+++ src/core/rovr-renderer/html-with-react.js
