+++ src/core/rovr-renderer/view.js
